Commands only update first matching config value.
bin/nesta can patch config.yml; this commit ensures that it will only
modify the first matching line it discovers in the config file. It used
to change them all, but we rarely want it to modify config settings in
production, so the settings that are defined later in the file should be
left well alone.