Moved Nesta::App.root to Nesta::Env.root.
If you want to load Nesta inside another Ruby app you'll probably want
to store it in a directory other than the process's current directory.
To do so you'll need to set the root directory, and you have to do that
before nesta/app is loaded. Given that nesta/app was what defined
Nesta::App, this seemed like a daft place to store the root directory
setting.