Added 'Articles heading' metadata.
By default (in the default theme) Nesta adds a heading above the
articles that are listed on a page, entitled "Articles on Thing" where
Thing is the name of the page that the articles are listed on. If you
set the 'Articles heading' metadata Nesta will allow you to override the
entire heading for a given page.

Only themes that call the articles_heading helper method will support
this metadata.