Remove parent page's heading from title tag.
See #108 if you're interested in reasons why this makes sense. To be
fair, specific titles (as supported by this commit) are better for your
users, and better for your SEO (should you care). If you want longer
titles with more contextual info in, you can still set the 'Title'
metadata on each page.