Add config key to customize default 'Read More'
The user is provided the config key `read_more` which will set the
default text used for 'Read more' links throughout the site. If the
`read_more` config value is not set, the site will use the previous
default of 'Continue reading'. Users can still customize the 'Read more'
text on a per-page basis with metadata.