Removed Travis build status from README.
The build usually fails due to the environment on Travis not coping for
some reason, so pointing out that Nesta's tests aren't passing isn't
accurate or useful.

It's currently failing on 1.9.3 because rubygems needs updating on
Travis. Ruby is such a frigging ghetto; rubygems should be far more
mature than this by now. But such is our lot...