Config.from_yaml should never return nil
When changing the protocol of from_environment and from_yaml such that
they should always raise Nesta::Config::NotDefined if they are unable to
return a configured value of a config setting, I'd introduced a bug.

If any NESTA_ environment variables were set, or if the config.yml file
didn't exist, from_yaml would return nil. This broke the behaviour of
other methods that depended on it to return NotDefined.