WIP: Stop creating too many <li> tags in nested menus
Rather than running round the `menu.each` loop in `display_menu`, we
need to use recursion to consume menu data structure. The problem is,
inside `display_menu_item` we need to be able to insert a inside the
<li> tag that we output the link text. To do that we need to be able to
re-call display_menu_item` from within `display_menu_item`.

Needs better tests, and a bit of a rewrite.