Add dependency on Tilt 1.4
Tilt 2.x will try to require the gems that implement the different
Markdown processors (e.g. Maruku), irrespective of whether or not it's
installed on the system. This can cause a LoadError.

It's new behaviour, and Nesta will need updating to avoid it. Requiring
version 1.4 works around it in the meantime.