Add heading_or_site_title helper method
Some themes include links for sharing pages on social networking or
bookmarking sites. They pass the page's heading to the third party
services, to serve as a description of the page being shared.

Sometimes it's possible to share a URL that isn't backed by a physical
page with a heading (such as a 404 page). Sometimes a page won't
actually have a heading.

In these cases, we can fall back to using the site's title to describe
the page. This new helper will return the appropriate text.