Upgrade to tilt 2.x
Some of the code in tilt 1.4 generates warnings when run under Ruby 2.4.

Here's an example warning:

    /path/to/.gem/ruby/2.4.2/gems/tilt-1.4.1/lib/tilt/haml.rb:55:in
    `block in precompiled_postamble':
    Haml::Engine#precompiled_method_return_value at
    /path/to/.rubies/ruby-2.4.2/lib/ruby/2.4.0/forwardable.rb:157
    forwarding to private method
    Haml::Compiler#precompiled_method_return_value

Upgrading to tilt 2.0 silences it.

However, tilt 2.0 has started automatically require'ing gems that are
used to process templates (e.g. `maruku`) as soon as the corresponding
template class (e.g. `Tilt::MarukuTemplate`) is referenced. This means
that referencing `*Template` classes for processors whose gems aren't
installed raises load errors.

I've therefore had to update the code that registers the Markdown
template handlers in `models.rb` so that it can ignore `LoadError`.
Note that the `rdiscount` gem is required explicitly in `models.rb` as
RDiscount is used by default.

I think it would be sensible to stop requiring `rdiscount` here, adding
each site's preferred markdown processor to its local `Gemfile` instead.
I'm not doing that now as it would be a breaking change, so is out of
scope of this commmit.

Were we to move to specifying the default Markdown processor in each
site's `Gemfile` we'd also need to check whether Haml's inline Markdown
feature works with other Markdown processors. I think it requires (or
used to require) `rdiscount`.