Fix current_item? helper when mounted at /prefix
As reported in issue #112:

> If you mount Nesta on a path other than / the `current_item?` helper
> fails as it compares the page's path relative to the `content/pages`
> folder with the value returned by `request.path`.

The solution was investigated on #112, fixed in a pull request (see #156),
but was never merged to master.

This commit ports the fix that @jkowens put together, updating the tests
to work with the new test suite (we've moved from rspec to minitest
since #156 was first submitted.