WIP: Setup Docker container for running tests
I'm not yet sure whether it'd make sense for Nesta to define a
Dockerfile that you can use to build an image in which you could then
run Nesta.

On the one hand, others who see the Dockerfile will probably expect that
it's a good way to build a container that can run a Nesta site.

On the other, creating docker config for a running site feels like it
belongs in the templating system that's called when you generate a
project for a new site.

Meanwhile, what I really want this Dockerfile for is to run the test
suite, which means that running it as root isn't a security hole and I
can get away with a really simple Dockerfile for that.

Run a command in the container (and via `bundle exec`) like this:

    $ docker-compose run --rm dev ruby -Itest test/unit/config_test.rb

The default command is `rake test`, so you can run the entire suite like
this:

    $ docker-compose run --rm dev

The .env file that I've added to .gitignore is used by docker-compose.
I'm using it to temporarily override the minitest reporter, as tests
that require git to be installed are failing inside the Docker
container.

I could add git to the container and configure it, but it doesn't feel
like the clean way to solve that problem. We've already had to do
something similar for Travis when replacing rspec with minitest (in
commit 95259978). It's a smell; I don't feel that Nesta's test suite
should need to actually call git.