Configure Travis to use bundler 1.x
You might be thinking "Why not use Bundler 2?"

Because Bundler 2 doesn't work on Ruby 2.2, and Nesta is still running
on some sites that haven't yet been upgraded to Ruby 2.3 or above.

So for the moment, Bundler 1.x is fine. Note that Travis still defaults
to using Bundler 1.

See this page in the Travis docs for more details on the commands added
in this patch:

https://docs.travis-ci.com/user/languages/ruby/#bundler-20