Add Ruby 3.0 to continuous integration build
This commit removes Ruby 2.5.x from the Travis build. This isn't because
I know it no longer works, but because I've decided to aim to support
the three most recent releases of Ruby.

Version 2.5 probably still works (though I haven't tested it with the
libraries that are updated in this commit).