Stop executing commands when running tests
Since it was ported to Minitest the test suite has been executing
commands (e.g. `git`). The `Command::run_process` function was easy to
stub out under RSpec, but it had to be executed by Minitest as it
doesn't allow such indiscriminate mocking or stubbing (this is a
feature, encouraging better design).

This refactor uses dependency injection to remove the commands'
dependency on `Nesta::Process`, allowing the tests to pass in any
object with a `#run` method. And as such, those objects no longer need
to run any external processes at all.

Instead it's enough for the tests to assert that the appropriate
arguments have been passed to them.

Note that the submodules in `test/fixtures` are now unnecessary, and
have also been removed. The accompanying change to the Travis config
reverts the change made in 95259978.