Rename Process class to SystemCommand
Calling the `#run` method on this object creates a Unix process. It
doesn't, therefore, make much sense for the object itself to *be* a
process. It's something else; something that can create a process.

We "run" system commands though, so I feel this is a better name for the
class.