Render Sass with Dart Sass instead of libsass
The sassc Ruby gem is built on top of the libsass library, which was
deprecated [1] back in October 2020.

The best approach moving forwards is to use the sass-embedded gem, which
uses the Dart Sass executable to convert Sass to CSS. This has been
supported by Tilt since 2.0.1, and all we need to do to enable Tilt's
support for it is to include `sass-embedded` in our bundle.

Note that Nesta now renders Sass files by calling Tilt directly, rather
than delegating to Sinatra to find the templates. We could still do that
as Sass is still supported in Sinatra 2.x, but Sass has been removed in
Sinatra 3.0.

I've therefore moved straight to using Tilt directly, in preparation for
upgrading Nesta to run on Sinatra 3.0.

Note that Dart Sass has moved on from the syntax that was supported by
libsass, and some stylesheets will need updating (see [2]).

This commit modifies a couple of the included styles in order to fix
warnings produced by Dart Sass (see [3]).

[1] https://sass-lang.com/blog/libsass-is-deprecated
[2] https://sass-lang.com/documentation/breaking-changes
[3] https://sass-lang.com/documentation/breaking-changes/slash-div