Merge Nesta::Overrides module into Nesta::View
Three of the helper methods in Nesta::Overrides::Renderers need
modifying, now that Sinatra no longer provides methods to render Sass
and SCSS files.

These methods previously called `super` to make use of Sinatra's support
for these template formats.

I'm about about to implement them directly, using an alternative
implementation of Sass. As such, they won't be "overriding" any of
Sinatra's built-in methods any more.

I was about to move them into a new module to reflect that, but they
have a lot of behaviour in common with the other template rendering
methods in the Overrides::Renderers module.

Surely, it makes more sense to keep them together, but to update the
name to reflect the fact that they don't all lean on Sinatra's built-in
behaviour. The fact they all called `super` was just an implementation
detail!