Replace custom code with Sinatra's #find_template
Nesta allows you to put your template (e.g. Haml, Erb, Sass and CSS)
files in a local `./views` folder, or in a folder inside a theme. It
also supports loading default files from within the gem itself.

Sinatra didn't originally support multiple template directories, so
Nesta contained custom functionality to make it work. The overridable
`#find_template` method was added to Sinatra, which makes it easier to
do. [1]

This was pointed out by @lilith back in 2012. It always looked sensible,
but it wasn't pressing. But I'm working on upgrading Nesta to Sinatra 3,
which has dropped support for Sass, so I need to refactor this code
anyway. So let's start that change from a better place. Thanks @lilith!

[1] http://www.sinatrarb.com/2011/03/03/sinatra-1.2.0.html