Add command for generating static sites
The build command will iterate over all the pages in your site,
converting each of them to HTML, saving them in a folder called
"./dist". If you want to save them to a different directory, specify it
as part as the command:

    $ bundle exec nesta build         # builds into ./dist
    $ bundle exec nesta build output  # builds into ./output

Note that we check to see whether or not the build directory has been
set to "public", in case somebody tries to use that folder by accident.
The "./public" folder is used for storing static assets (CSS, JS, images,
etc), so it wouldn't be wise to overload that.

I've explicitly set Sinatra's public_folder setting to "public" (in
`app.rb`) so that we can use it as part of this check (the default value
is `nil`, which means "public" but isn't a particularly helpful
default!). This means that if you really want to build your site into
the ./public folder you've got the option to move your ./public folder
out of the way, then set the public_folder setting accordingly in your
site's `app.rb` file.