Set Sinatra's root setting for static build
The root setting is used for rendering some pages, and for determining
some paths. When Nesta runs as a normal Sinatra app the root directory
is set within config.ru, before the Sinatra app gets created.

The build command makes the assumption that the build command is being
run from within the project's top-level directory, which it uses to
determine the value of Sinatra's root setting.