Build 404 page for static sites
Both Netlify and Vercel default to reading a static site's 404 page
from a file called `404.html` in the root of the build folder.

I'm therefore considering this as a convention, and have hard coded the
path.

Note that the log message that announces that a page has been rendered
is now guarded by a check to see whether the file contents have changed.

We don't know which template (or app code) the 404.html page was created
from, so we're unable to check whether or not it's out of date. So we
re-generate 404.html on every run. Printing a message to say that we've
done so even if the source files haven't changed could be confusing.
Hence the check to see if the contents have changed.