Build static copy of XML sitemap
This code will fail if the sitemap isn't available at `/sitemap.xml`
(which is where Nesta puts it if you don't override it).

It's possible that somebody will be serving their sitemap from a URL
other than `/sitemap.xml`, but in the spirit of shipping something
simple and iterating on it, I'm assuming people will have stuck with the
default location.