Use singleton to tidy up Nesta::Config
We care about only having one instance of the config object so that we
only pay the cost of reading the YAML file once.

It's a long time since I've used the Singleon design pattern, but in
this case it's an improvement.

All the code for accessing the contents of `config.yml` was implemented
as class methods, which is effectively no different to having a global
variable. These class methods led to the code being more complicated than
it otherwise needed to be.

In order to simplify it I've converted the logic to instance methods.
The class-level API (e.g. `Nesta::Config.title`) has been maintained by
forwarding calls to the class methods to the single instance, which is
achieved via the inclusino of `Forwardable` and use of `def_delegators`.

The `Singleton` module automatically ensures that there's only one
instance of the `Config` object, which is accessible via the `#instance`
method.

Note that I have made a potentially breaking change here, if anybody was
relying on the existence of `Nesta::Config.yaml_path`. That's not a
great name, and it's arguably not in the right place. So I've moved it
to `Nesta::ConfigFile.path`.

I've also removed the unused `author_settings` class variable, that
somehow managed to hang around (unused) for about 14 years!