Set site's domain name for use in static builds
Some URLs need to contain the domain name of the web site, such as XML
sitemaps and Atom feeds.

The domain name can't be determined at run time for a statically
generated file, so we need to specify it at build time.

The `nesta build` command now takes an optional argument:

    nesta build --domain mysite.com

You can also specify it in `config.yml`, like this:

    domain: mysite.com