Build static copy of Atom feed
This code will fail if the Atom feed isn't available at `/articles.xml`
(which is where Nesta puts it if you don't override it).

It's possible that somebody will be serving their feed from a URL
other than `/articles.xml`, but in the spirit of shipping something
simple and iterating on it, I'm assuming people will have stuck with the
default location. For now.