Rename SiteContent to Site
I'm about to add some new behaviour to this class that doesn't have
anything to do with the site's content.

We need to be able to convert Sass and SCSS to CSS, along with a range
of other templated assets. They're not related to the site "content",
but are still part of the site.

It feels sensible to ask the same object to build them; its responsibility
is to use Nesta's Rack app to convert a source file into the "compiled"
output that we will serve to the browser.

So `Static::Site` feels like a reasonable name for it.