Build static copy of templated assets
Plenty of the files delivered to the browser are converted by Nesta.
e.g. Sass or Less are converted to CSS. It's done automatically at
request time when Nesta is running behind a web server, but we need to
explicitly generate and include these files included in the build.

Nesta has no (easy) way of knowing which of the templates in the views
folder need converting like this, so we list them in the config file.

Nesta will then request them all at build time, and save the results
into the build directory.