Rename CHANGES to CHANGELOG.md
The `.md` file extension causes GitHub to render the file as Markdown.
I've made a few tweaks so that'll work out nicely.

I've renamed it to CHANGELOG to reflect the current fashion (I think
CHANGES is what some parts of the Unix community called them a couple of
decades back).