Drop support for Ruby 2.7
Nesta's code runs fine under Ruby 2.7 (and most likely earlier), but I'm
dropping support for 2.7 as Nokogiri (one of Nesta's dependencies) now
requires Ruby 3.0 or later.

We're now in a position where Nesta's dependencies can't be installed
since the recent security patch released in Nokogiri 1.16.2.

I'd consider not upgrading Nokogiri for a while, but its bundled version
of libxml2 is responsible for most of the security vulnerabilities that
crop up in Nesta.

So sadly, support for Ruby 2.x has come to an end.