Support for frozen string literals
When run with the --frozen_string_literal flag, Ruby will throw an error
if you try to modify a string object.

Leaving discussions of whether this approach is idiomatic Ruby (or not)
to one side, it seems sensible in light of the discussion in [1] and [2]
for projects like Nesta to work without mutating strings. And note that
in Ruby 3.4, things are changing.

To find the areas of the code that warranted an update I ran the tests
under Ruby 3.3, like this:

    RUBYOPT=--enable=frozen-string-literal rake test

[1] https://bugs.ruby-lang.org/issues/16047
[2] https://bugs.ruby-lang.org/issues/20205