Update Sinatra to fix vulnerability
See https://github.com/gma/nesta/security/dependabot/47 for details.

Note that I've also pinned tilt to a version less than 2.5.0. This is
because 2.5.0 removed support for the deprecated Maruku markdown
processor.

Haml versions lower than 6.0 attempt to register Maruku handlers, which
throws an exception that we're unable to catch.

Supporting Haml 6.0 or above is outside the scope of this commit, so
we'll be restricted to versions of Tilt that support Maruku for now.